requirements.txt
PyYAML
aiogram
pyrogram
SQLAlchemy
psycopg2-binary
chromadb-client
mistral-client
sentence-transformers
torch
